Guzzle receives `RequestInterface` on send* method
Currently, code is sending `ServerRequestInterface` instance (which extends
`RequestInterface`) to Guzzle. Nothing breaks, apparently, but Guzzle isn't
aware of extra methods from `ServerRequestInterface`, meaning whatever it's
set which is not on `RequestInterface` is completely ignored.

This PR intends to change `ServerRequest` to `Request`, and properly add
query parameters and cookies (considering `RequestInterface` lacks methods
such as `withQueryParameters` and `withCookies`).